pyproject.toml
[project]
name = "django-starter-pack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "channels-redis>=4.2.1",
    "daphne>=4.2.0",
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "huey>=2.5.3",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.11",
]
